Explicitly cast to int to get around illegal array key type float warning

+++ app/helpers.php
@@ -102,7 +102,7 @@ function format_bytes($size, $precision = 2)
     $base = log($size) / log(1024);
     $suffixes = array('', 'k', 'M', 'G', 'T');
 
-    return round(pow(1024, $base - floor($base)), $precision).$suffixes[floor($base)];
+    return round(pow(1024, $base - floor($base)), $precision).$suffixes[(int)floor($base)];
 }
 
 function get_host_from_url($url)
